New DB schema. Old data migrated. Replay Added.
This compromises 3 weeks of work to add the new event-based database schema,
port the entire codebase to use it, migrate the old data, and then validate the
data. Replay was fixed and used to validate the migrated data.

The new event-based database merges many of the previously existing tables.
Each game is now represented by a linear list of events, each with a timestamp,
role, origin, and several other key parameters. This will make writing scripts
to work with the data far easier, and it will enable future optimizations
simpler. This includes improved DB buffering for better ammortized disk
performance.

The replay mechanism has been rewritten from scratch to depend on the
event-based database (previously, it just replayed recorded network messages).
This reduces the amount of storage needed, as we no longer need to record
network messages. We also don't need the recorded messages, as we've been using
the current code for a while and haven't had any major issues with what is
recorded to the database for a long time.