Adds support for scenario rooms.
New lobby by default in config: "Scenario-lobby". Rooms in this lobby are
1-player lobby rooms. The player plays as a follower, and scenarios can be
loaded in the Unity UI via the escape menu.

In all non-mturk and non-google games, state downloading has been enabled in the
UI. Hit the escape key and click "Save Scenario State". Then, when in a scenario room, you
can hit "Upload Scenario State" from the escape menu to resume.

The scenario state file is a pretty-printed JSON file. The schema is defined by
the python dataclass:

@dataclass(frozen=True)
class Scenario(DataClassJSONMixin):
    map: MapUpdate
    prop_update: PropUpdate
    turn_state: TurnState
    objectives: List[ObjectiveMessage]
    actor_state: StateSync

This is defined in server/messages/scenario.py.