Client exception uploading.
This allows us to capture and save exceptions that happen in the client. This is
to-date the source of many of our unfound bugs. When an exception occurs, the
stack trace and all client-side module logs are uploaded to the server, where
they are stored in a database.

This commit is a work-in-progress. Expect further testing soon.